integer to double in player rating
Changed integer value of rating arays data type so that it can hold non integer rates as well

+++ backend/app/models/player.model.js
@@ -13,7 +13,7 @@ module.exports = (sequelize, Sequelize) => {
         type: Sequelize.INTEGER
       },
       p_rating:{
-        type: DataTypes.ARRAY(Sequelize.INTEGER)//0:total rating 1:total fprating 2:#ofratings
+        type: DataTypes.ARRAY(Sequelize. DOUBLE)//0:total rating 1:total fprating 2:#ofratings
       },
       position_a: {
         type: Sequelize.STRING
